New address data type

+++ include/opcodes.h
@@ -37,7 +37,7 @@ typedef double cpuNumber_t;
 typedef union
 {
     cpuNumber_t number;
-    size_t address;
+    long long int address;
 } cpuData_t;
